log: move mock assets to separate file
