+++ log/mock_test.go
@@ -1,0 +1,11 @@
+package log
+
+import "testing"
+
+func TestNewTestLogger(t *testing.T) {
+	logger, out := NewTestLogger()
+	logger.Info("hi")
+	if out.All()[0].Message != "hi" {
+		t.Error("bad logger")
+	}
+}
